feat: Added PublicationsPaginatedList  on AnnouncementsPage
